feat: support getting profiles from execute state
